Bump to 120.0.6099.71-1
We cleaned unused patches and synced PKGBUILD with extra/chromium. So
there are so many files got removed.

See https://build.opensuse.org/package/show/home:ZhangHua/ungoogled-chromium-xdg
to check build log on OBS, but please note that we changed PKGBUILD a
bit to meet OBS's requirements.

For example, `java-runtime-headless` is required
to be pinned to an exact package so I use
`jre-openjdk-headless` instead. What's more,
`depends` is not allowed to be modified in
runtime, it must be defined directly so OBS can
find which packages are needed to be installed,
so I commented line 100 and set them in line 25
manually.